Updated LoadOmnicMAP module to work with data collected with an Array detector
